Upgraded the project by inserting it line by line into a clean Altirra 4.40 source tree. No other changes were made.

+++ src/ATCore/source/sioutils.cpp
@@ -21,9 +21,9 @@
 #include <stdafx.h>
 #include <at/atcore/sioutils.h>
 
-uint8 ATComputeSIOChecksum(const uint8 *p, int len) {
+uint8 ATComputeSIOChecksum(const uint8 *p, size_t len) {
 	uint32 checksum = 0;
-	for(int i=0; i<len; ++i) {
+	for(size_t i=0; i<len; ++i) {
 		checksum += p[i];
 		checksum += (checksum >> 8);
 		checksum &= 0xff;
